backend/requirements.txt
# ── AI Drive Agent — Python Dependencies ──────────────────────────
# Install with:  pip install -r requirements.txt

# Web framework
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.34.0,<1.0.0

# Google APIs
google-auth>=2.38.0,<3.0.0
google-auth-oauthlib>=1.2.1,<2.0.0
google-api-python-client>=2.159.0,<3.0.0

# Gemini LLM
google-generativeai>=0.8.0,<1.0.0

# LangChain agent framework
langchain>=0.3.14,<1.0.0
langchain-google-genai>=2.0.9,<3.0.0
langchain-core>=0.3.29,<1.0.0

# Security
cryptography>=44.0.0,<45.0.0
python-jose[cryptography]>=3.3.0,<4.0.0

# Configuration
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.7.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0

# HTTP client
httpx>=0.28.0,<1.0.0
